(NPUP-15) Port modulo operator to the dispatcher.
Porting the modulo operator's implementation to the dispatcher.
