+++ lib/include/puppet/compiler/evaluation/operators/binary/modulo.hpp
@@ -1,0 +1,23 @@
+/**
+ * @file
+ * Declares the modulo operator.
+ */
+#pragma once
+
+#include "descriptor.hpp"
+
+namespace puppet { namespace compiler { namespace evaluation { namespace operators { namespace binary {
+
+    /**
+     * Implements the modulo operator.
+     */
+    struct modulo
+    {
+        /**
+         * Create a binary operator descriptor.
+         * @return Returns the binary operator descriptor representing this operator.
+         */
+        static descriptor create_descriptor();
+    };
+
+}}}}}  // namespace puppet::compiler::evaluation::operators::binary
